requirements.txt
Flask==3.1.2
netmiko==4.6.0
python-dotenv==1.1.1
